Implement a Simple depot based on a merkel tree implementation
This implementation regrettably breaks a lot of tests (namely Resolver
and Engine) because the Rehydrate method of the new resolver does not
know when to break the event consumption channel range loop.

My likely fix is to start tracking the HEAD pointer on the refdb on the
SimplePartition- and SimpleEventIterators so that for any given even
I know if this is HEAD or not. I could use the same attribute to track
whether a new head is a decentant of the old head (fast forward) or
whether the HEAD pointer has jumped which the Iterators need to know
anyway, in order to know whether to restart, or emit new things (once
they have a mechanism for subscribing to changes in their underlying ref
and objectdb stores contents).