Do not start a go routine for handling _each message_
The partition handler does have to be backgrounded with `go …` however
the events _must_ be handled in the order in which they are received.

By using the go keyword to start a goroutine for every message the
number of goroutines blew up very quickly (to a small number in this
small test example) and left the determinism of the test at the mercy of
the Go scheduler.

By allowing "backpressure" of sorts to be applied to the producer and
handling the messages synchronously a test run with `-count 10000` ran
successfully with and without the race detector. 🎉