Implement Partition subscription
This commit introduces the idea of a partition iterator emitting new
partitions on head pointer movements.

The depot.Glob() method now tracks a list of subscribers (implemented
using channels) as they are emitted.

Changes to the head pointer are propagated with a RefMove object with
two fields `from` and `to` of type Hash. A 3rd field is tracked a
boolean indicating whether this is a "fast forward" pointer move. This
code path is not implemented and the field always carries the zero
value.

Changes to existing tests were necessary as the previous "everything in
the correct order test" relied on the partitionIterators channel
eventually being closed so that it could continue to an early exit
condition.

The refactored test reads more cleanly and passes with or without the
new feature code.

The subscription mechanism may leave a lot to be desired. Subscribers
are notified sequantially, in a blocking fashion with a very high
timeout and inelegant error handling.

A head pointer move does not honor the "from" field of the RefMove
struct and will always walk the checkpoints until the first checkpoint
on this branch with no parent (effectively walking back to the first
checkpoint every time) this should be easy to fix, along with the code
to check for ff/non-ff head pointer movements.

The next step will be to make event subscription work and then to revist
the ovservability and performance side of things.