Optimisation for walking relevant checkpoints
This ensures we never walk too far back in the history.

Given checkpoints A, B, C, D, E, F prior to this commit any subscriber
would see A..F on the stack even if the head pointer moved only E..F. A
caching mechanism ensured that partition iterators _already_ emitted in
A..E would not be emitted again, so only event emitters for new
partitions would be emitted.

With this commit in place the subscriber receives a partial stack
containing only the contents of E..F. The deduplication mechanism is
still valid incase E..F contains new _events_ for an existing
_partition_.

With this change in place PartitionIterators can notify their
EventIterators of new events by passing them the stack over a channel
and allowing the event emitter to unpack the related affixes and events
and handle them accordingly.