Introduce event subscriptions
With this commit consumers having made a depot.Glob() call will not only
receive new _partitions_ when head pointers are moved to a forward
position including previously un-broadcast partitions, but on existing
partition subscriptions (EventIterators) they will also receive new
_Events_.

In summary, with the code in this commit a reading client can be assured
to get all existing and future data as long as they remain connected.

Branches are still not handled. (Checkpoint.ParentHashes[0] is
referenced rather than a more clever method merging histories).

This commit improves the pattern matching and linked-list walking to
gather relevant checkpoints to respect the from, and to type.Hash
arguments meaning that a head pointer move from E..F will walk no
further back in the history than E and will emit a stack of depth 1.

This was necessary to avoid emitting already emitted events.

A reasonable next step would be to handle the non-ff case (which should
likely always trigger an A...F stack walk, or to look at pulling the
depot out into a networkable service with net/rpc or protobufs over
GRPC.