Remove .Event() from PersistedEvent
Prior to this patch a PersistedEvent (something returned by the Depot)
could resolve, and rehydrate an Event. Thsi was implemented early in the
Depot design as it seemed like a logical fit that an event name, and an
encoded payload should be able to be turned into a real typed instance
of an Event.

In order to facilitate this, the Depot carried a reference to an
EventRegistry which had two major downsides:

- tests in framework/depot were reaching out to ../../event (out of the
framework package) to get hold of an event registry instance.
- if a Depot would be offered as a networked service it must share the
same runtime definitions as the Engine which is another level of
concern

The first point could have been addressed by bringing the EventRegistry
into the ./framework/ package but that seemed like a mixing of concerns.

The second one however becomes very problematic now that the Depot has a
method based interface that lends itself well to being handled over a
network.

Indeed the next step may be to write a GRPC server/client for Depot and
make sure that the concepts translate well.

The tests have become somewhat less accurate after this change, without
the ability to deserialize and really _type_ a retreived event at the
Depot level, a string comparison would be the next-best on the payload
carried within the event. This may something worth adding back in soon.