Expand demo server (listing, centralization)
This commit rounds-out the demo server with the following features:

- Listing of Commands, Events and Aggregates (untested, likely subject
to change)

- Integration of the previously segregated "Depot server" (which is
really just a HTTP server over the depot storage layer, it doesn't
actually respect Depot semantics)

- Building and live-reloading with Gin which was simple but mandated the
use of flags passed as env vars not only as real flags. Important for
the storage path.

With this commit in place the following is possible:

    [~]$ curl localhost:3000/list/commands
    {
        "aggregates.Session": [
            "*Start"
        ],
        "aggregates.WidgetsApp": [
            "*AllowCreationOfNewIdentities"
        ]
    }

    [~]$ curl localhost:3000/list/events
    {
        "AllowCreateIdentities": {},
        "DisableCreateIdentities": {},
        "SetAvatar": {
            "contentType": "",
            "imgData": null
        },
        "SetDisplayName": {
            "name": ""
        },
        "StartSession": {}
    }

    [~]$ curl localhost:3000/list/aggregates
    {
        "_": "aggregates.WidgetsApp",
        "session": "aggregates.Session"
    }

And the previously possible:

    [~]$ curl localhost:3000/ref/
    {"depot_common_test519643404/refs/heads/master":"sha256:5a7d7d9db3bbf8f453a96df99d9cfe1043ef09aa34c44a3171e2ba971387e17f"}
    [~]$ curl localhost:3000/obj/sha256:5a7d7d9db3bbf8f453a96df99d9cfe1043ef09aa34c44a3171e2ba971387e17f
    {"affixHash":"sha256:5c726b8009ee9d81dba151d70397c4f6dbbeecef8195b0e2b79b8b9fe97a0fd9","parentHashes":null,"fields":{"session":"hello world"},"summary":"","commandDesc":"eyJzdHViIjoiYXJ0aWNsZSJ9"}

The APIs and implementations are inconsistent, and the obj and ref are
probably better references than the untested spikes to list registered
entries in the three manifest types.

This server also responds to posted commands (the body is currently
ignored).

Following this change the next step will be to make sure that at least
one meaningful command and some meaningful rehydration pathways for
aggregates are written and that the Engine does the right things with
the result values returned by Commands.

As of this commit the Engine is making changes to the object and ref DBs
but not in a way that makes sense to me, and is almost certainly just
accidental vandalism.