General clean-up and docs.
This commit changes the signature of Command.Apply to take a
types.Session in the second position which is a type alias for
types.Aggregate, this is much more descriptive.

Additionally many more types have doc strings and some redundant types
have been removed.

Redis Depot which conformed to a previous version of the API has been
removed it should be trivial to add it back in the future. (The API
predated the simplification to use an object- and ref-db and wrap them
in a type carrying the query logic.