Implement NamedAggregate
In order for Commands to be able to properly populate a
types.CommandResult the aggregates need to track their own name when
crossing the boundary from Engine->Command-Engine.

Aggregates are named only by their partition name and don't
intrinsically carry an ID.

This change ensures that an Aggregate which is retrieved from the Engine
(rehydrated) will be given an ID at the right time and that this ID will
propagate to the Command and through the return value of Apply
(types.CommandResult). The Command need not know the identity of the
Aggregate unless it has a meaning for the business rules.

This work as it stands now is incomplete. There's work to be done to
figure out how to address things like creating a new user, where the
{"path":...} on the command description may be empty, or simply "Users"
and there isn't really an Aggregate instance to which to direct the
"new".