Milestone Engine work for usability.
With this patch the Engine may finally be ready for use.

The following improvements have been made:

- Engine now creates checkpoints with the given []byte containing the
command description (the argument to Apply)

- Engine now tracks, and respects the position of the Head Pointer (if
any) when running Apply and StartSession

- Engine now uses the NamedAggregate when persisting events to build the
Affixes.

The general outcome is that now Engine actually respects user input and
should not vandalize history by forcing a head pointer move to a new
orphaned checkpoint.

There are many other supporting changes:

- The Depot interface is widened to include a HeadPointer() method to
return the head pointer (if any) it may return nil, nil if there is no
head pointer yet on the current branch.

- Depot.DumpAll is avaialble via an interface upgrade to Depot. It
compares a simple text representation of the Depot which makes
comparisons relatively easy (using cmp.Diff) however a proper diff
mechanism for Depot which respected spitit of what equality means for
Depots wouldmake the tests more readable and leave scope for better
error messages.

- The use of an PRNG in the tests for session IDs was dropped. In
-count=1 test runs this worked OK, however with -count=>1 the PRNG
was not re-seeding, this also led to problems when running the whole
suite, and not only a single named test, this was further exacerbated
by the .Parallel() helper on the tests which made tracking this down
difficult.