Fix Command and Aggregate tests
These are usually excluded from the test suite as the Makefile
hard-codes PACKAGE=...framework/... which is a level deeper than the
aggregates, events and commands are stored.

With this patch everything passes cleanly but it is somewhat misleading
as the Command tests use an empty Depot and the EventFixture type in the
types package is insufficient.

Committing here as there is a larger refactoring to raise the
.Rehydrate() method out of the depot. As-is the Depot has no
EventManifest and can not realistically ever take this responsibility as
it should live out of process.

The next logical refactoring would be to pull the rehydration out of the
depot and pass something other than a Depot to the Commands.