Implement and mount CreateIdentities command.
This needs some love (nothing to stop me making dupes) - but the time is
now to address how targetting anonymous aggregates should work to create
them for the first time.

This is an OK solution for the time being, but needs more testing before
I can build out any app logic on it.