Refactor Depot <> Repository
DDD names a "Repository" as the place that entities live. With Retro
that was blurry until now.

After this commit the Repository is used in all Commands and the Engine
takes a Repository and a Depot; the former concerns itself with all read
operations and the latter with all write operations.

The repository/simple.Rehydrate is misbehaving presently but this can be
tested for and fixed soon. It currently replays (apparently) all events
for all aggregates in related affixes, but it's not 100% clear.