Introduce inflection for Event names and Commands.
With this patch the Engine and *Manifest types now understand string
inflection.

Two are in wide use:

  - PascalCaseForCommandNames
  - underscore_names_for_everything_else

Prior to this patch Events would be named per the Go type naming rules,
so Affixes would be written with event names such as `SetDisplayName
{...}`, they are now written with underscoreized names e.g
`set_display_name`. The EventManifest takes care of the details.

The resolver has also been modified to accept either `PascalCase`
command names or `underscore_case`, the following are both valid, and
route to the same Command:

  {"name":"create_identity", "path": "identity/lee6", "args":{ .. }}
  {"name":"CreateIdentity", "path": "identity/lee6", "args":{ .. }}

The patch also makes minor changes to the demo server listings of
commands, events and aggregates to respect the new naming, and hide some
of the Go type system artifacts which were leaking across the HTTP
boundary.