Changes to the naming of things.
Prior to this patch a non-existent aggregate could be targeted by name
via the resolver (such as in a command description). This was convenient
but blurred the line between persisted and unpersisted aggregates and
felt like a code-smell.

With this patch in place the resolver will err if asked to resolve
something like {"name":"foo", "path":"does/notexist", ...}.

The "correct" way (for now) to use anonymous aggregates:

- Mount the command a level higher, on an existing aggregate (such as
the root) which takes ownership for the creation of new things. For
example "CreateIdentity" command has moved from being mounted against
the Identity Aggregate and is now mounted on the root aggregate ("_").

- The re-mounted aggregate can simply reference a new aggregate:

      return types.CommandResult{
        session: []types.Event{
          &events.AssociateIdentity{Identity: cmd.identity},
        },
        &aggregates.Identity{}: ownEvents,
      }, nil

  This command result adds one event to the existing session (received
  in the args) and creates a new &aggregates.Identity{} which has a set
  of it's own events.

This change was fairly invasive, calls to CreateIdentity from outside
now look like this:

{"name":"create_identity", "args":{"name": "Lee H", "publiclyVisible": false, "avatar":"R0lGODlhAQABAIAAAP///wAAACwAAAAAAQABAAACAkQBADs="}}

(*Note:* No "path")

The anonymous identity will be given a name in the persistence layer of
the Engine using the given IDFactory.