Refactor Resolver to return Command not CommandFunc
This allows broadening the Command interface to expose a potential
Render() function which the engine can call out of band to write the
result.

Through optional interface upgrade a Command can implement a rendering
function which can access the session, modified command result (with
Ids, etc from the persistence layer) and more.