Implement deep-naming of all referenced Aggregates in CommandResult
The key thing to note in this patch is that any retro.NamedAggregate
(retro.ExistingAggregate) which has _no name_ will be given one whether
on the key, or values (in the []retro.Event slice) side of the
retro.CommandResult object returned by a Command.

The use-case is simple, when creating a new Aggregate and referencing
it:

    var newIdentity = &aggregates.Identity{}
    // ... snip ...
    return retro.CommandResult{
      session:     []retro.Event{
        &events.AssociateIdentity{Identity: newIdentity}
      },
      newIdentity: []retro.Event{
        &events.SetAvatar{ImgData: cmd.args.Avatar},
        events.SetDisplayName{Name: cmd.args.Name},
      },
    }, nil

Whichever order the Engine chooses to persist the Evs, it will first do
a deep walk of all keys and values in the result map, and for the values
it will walk the types, checking if any of the fields of those event
types refer to something for which a name can be generated.

This ensures that whether the aggregates.Session{} or
aggregates.Identity{} is persisted first the names will be expanded such
that they serialize in a useful way.