Begin using golang.org/x/xerrors in JsonPacker and Depot
I have settled on the following schema for wrapping errors:

  return "foo", xerrors.Errorf("json-packer: %s: %w", err, ErrAffixScan)

Such a schema has the benefit of communicating the locality of the error
(json-packer) as well as the original error string (%s, err) and a
semantically useful error type `ErrAffixScan` which can be used higher
up the call-stack with xerror.Is(e, ErrAffixScan) to determine whether
the error confirms to that "type".

Placement of the : is only important in the case of `: %w`, from the
docs:

    If the last argument is an error the returned error's Format method will
    return it if the format string ends with ": %s", ": %v", or ": %w". If
    the last argument is an error and the format string ends with ": %w",
    the returned error implements Wrapper with an Unwrap method returning
    it.

In short the suffixes `: %s`, `": %v` and `: %w` all gracefully format,
but only %w semantically wraps the error in a way that .Is() or .As()
can determine the original type.