Refactor matchers to be stateful.
The DoesMatch() method previouly expected two arguments which made the
matcher "stateless". This refactor allows a matcher to carry some
internal state so callers need only pass the data they have at hand
without the callers tracking the matchers pattern semantics.

This patch opens the door for a more generic interface where a matcher
can receive a Checkpoint, Affix or Event and not only a
string/PartitionName.

A more generic interface comes with some complexity overheads but begins
to permit a more generic query interface where matches can be made
against metadata fields in checkpoint headers or potentially affixes
exclusively containing one or more than one entity.

Querying based on fields in events is also not completely out of the
question in this model.

The next logical refactor is to make the GlobMatcher here into a
"PartitionNameGlobMatcher" and have it take Checkpoints and Affixes, but
emit events via the PartitionIterator as it does presently.