Implement basic recursion in queryable struct
Demonstrates the simplest case of ranging over parents and collecting
more and more and more matches.

An alwaysMatches helper has been added to the tests to allow me to test
that in a naive case we'll always match everything ensuring we do a
proper full walk of the storage.

Test setup had a mistake too as the head ref pointed at a checkpoint
which referenced an affix which doesn't exist, easily corrected.