Implement chronological indexer
This genericise the behaviour of collecting a historical record from a
given ref backwards and linerarizing the result.

The implementation is sub optimal, with a full tree walk from ref to the
first ancestorless checkpoint. Sometimes twice, if a merge checkpoint
is present in the history.

Worse, for each checkpoint discovered the list of all hitherto
discovered checkpoints will be binary-searched to find the correct index
at which to insert the incoming checkpoint.

Aside from practical concerns about realworld performance, the
implementation itself leaves a little to be desired. Waitgroups,
blocking returns, and conditional branching seem like I may have missed
a more elegant solution in the Chronological#run() function.

One noteworthy remark: we can reach "the beginning" 1+n times given no
cyclic histories. Possibly some of this book-keeping can be moved to the
tocCPs type where the slice can maybe converted to a single-use sized,
and buffered channel. This would eliminate a lot of the book-keeping
from the Chronological index.